Chore: Move Slack notifier to `notify` module
